feature(homePage): individual files for objects and animation for plants
