+++ spelling-bubbles/ContentView.swift
@@ -26,8 +26,9 @@ struct ContentView: View {
     @State var showingPauseMenu = true
     
     var body: some View {
-        GameView()
+        //GameView()
         //PauseMenuView()
+        HomePageView()
             .ignoresSafeArea()
     }
